+++ nilearn/input_data/tests/test_nifti_labels_masker.py
@@ -11,9 +11,10 @@ import nibabel
 import pytest
 
 from nilearn.input_data.nifti_labels_masker import NiftiLabelsMasker
+from nilearn.input_data import NiftiMasker
 from nilearn._utils import testing, as_ndarray, data_gen
 from nilearn._utils.exceptions import DimensionError
-from nilearn.image import get_data
+from nilearn.image import get_data, new_img_like
 
 
 def generate_random_img(shape, length=1, affine=np.eye(4),
@@ -348,3 +349,20 @@ def test_standardization():
                                    (unstandarized_label_signals /
                                     unstandarized_label_signals.mean(0) *
                                     100 - 100))
+
+
+def test_nifti_labels_masker_with_mask():
+    shape = (13, 11, 12)
+    affine = np.eye(4)
+    fmri_img, mask_img = generate_random_img(shape, affine=affine, length=3)
+    labels_img = data_gen.generate_labeled_regions(shape, affine=affine,
+                                                   n_regions=7)
+    masker = NiftiLabelsMasker(
+        labels_img, resampling_target=None, mask_img=mask_img)
+    signals = masker.fit().transform(fmri_img)
+    bg_masker = NiftiMasker(mask_img).fit()
+    masked_labels = bg_masker.inverse_transform(bg_masker.transform(labels_img))
+    masked_masker = NiftiLabelsMasker(
+        masked_labels, resampling_target=None, mask_img=mask_img)
+    masked_signals = masked_masker.fit().transform(fmri_img)
+    assert np.allclose(signals, masked_signals)