+++ doc/whats_new.rst
@@ -17,6 +17,10 @@ NEW
   These two objects also leverage the `NiftiMaskers` to provide a direct
   interface with the Nifti files on disk.
 
+Fixes
+-----
+
+- :class:`nilearn.input_data.NiftiLabelsMasker` no longer ignores its `mask_img`
 
 0.6.2
 ======